Move Trie definition together with "merging" History
